Extract HttpClientException to another file
